backend/requirements.txt
fastapi==0.104.1
catboost==1.2.1
pandas==2.0.0
numpy==1.24.2
lightgbm==4.0.0
optuna==3.2.0
requests==2.31.0
scikit-learn==1.2.2
uvicorn==0.24.0.post1
joblib==1.3.0
pyyaml==6.0
pyyaml~=6.0
transliterate==1.10.2
python-multipart==0.0.6

